+++ src/Voucherify.Client.net35.Example/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Voucherify.Client;
 
 namespace Voucherify.Client.net35.Example
 {
@@ -8,23 +7,18 @@ namespace Voucherify.Client.net35.Example
         private const string app_token = "<your_app_token>";
         private const string app_id = "<your_app_id>";
 
-        private async static void Run()
+        private static void Run()
         {
-            try
-            {
-                Console.WriteLine("Voucherify. Get single voucher.");
+            Console.WriteLine("Voucherify. Get single voucher.");
+            VoucherifyClient client = new VoucherifyClient(app_id, app_token).WithSSL();
 
-                VoucherifyClient client = new VoucherifyClient(app_id, app_token).WithSSL();
-                string voucherCode = "<voucher code>";
-
-                DataModel.Voucher voucher = await client.Vouchers.Get(voucherCode);
-
-                Console.WriteLine("Voucher. Code: {0} Camplaing: {1} Category: {2} Discount: {3}", voucher.Code, voucher.Campaign, voucher.Category, voucher.Discount.Type.ToString());
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("Exception occured: {0}", e);
-            }
+            client.Vouchers.Get("<your_voucher_code>")
+                .Then((voucher) => {
+                    Console.WriteLine("Voucher. Code: {0} Camplaing: {1} Category: {2} Discount: {3}", voucher.Code, voucher.Campaign, voucher.Category, voucher.Discount.Type.ToString());
+                })
+                .Catch((exception) => {
+                    Console.WriteLine("Exception occured: {0}", exception);
+                });
         }
 
         static void Main(string[] args)