+++ src/Voucherify.Client/Api/Vouchers.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 using Voucherify.Client.Extensions;
+using RSG;
 
 namespace Voucherify.Client.Api
 {
@@ -19,58 +19,58 @@ namespace Voucherify.Client.Api
             this.client = new ApiClient(voucherify);
         }
 
-        public async Task<IList<DataModel.Voucher>> ListVouchers(DataModel.VouchersFilter filter)
+        public IPromise<IList<DataModel.Voucher>> ListVouchers(DataModel.VouchersFilter filter)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder("/vouchers/").WithQuery(filter);
-            return await this.client.DoGetRequest<IList<DataModel.Voucher>>(uriBuilder.Uri);
+            return this.client.DoGetRequest<IList<DataModel.Voucher>>(uriBuilder.Uri);
         }
 
-        public async Task<DataModel.Voucher> Get(string code)
+        public IPromise<DataModel.Voucher> Get(string code)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder(string.Format("/vouchers/{0}", code));
-            return await this.client.DoGetRequest<DataModel.Voucher>(uriBuilder.Uri);
+            return this.client.DoGetRequest<DataModel.Voucher>(uriBuilder.Uri);
         }
         
-        public async Task<DataModel.Voucher> CreateVoucher(DataModel.Voucher voucher)
+        public IPromise<DataModel.Voucher> CreateVoucher(DataModel.Voucher voucher)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder("/vouchers/");
-            return await this.client.DoPostRequest<DataModel.Voucher, DataModel.Voucher>(uriBuilder.Uri, voucher);
+            return this.client.DoPostRequest<DataModel.Voucher, DataModel.Voucher>(uriBuilder.Uri, voucher);
         }
 
-        public async Task<DataModel.Voucher> CreateVoucherWithCode(string code, DataModel.Voucher voucher)
+        public IPromise<DataModel.Voucher> CreateVoucherWithCode(string code, DataModel.Voucher voucher)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder(string.Format("/vouchers/{0}", code));
-            return await this.client.DoPostRequest<DataModel.Voucher, DataModel.Voucher>(uriBuilder.Uri, voucher);
+            return this.client.DoPostRequest<DataModel.Voucher, DataModel.Voucher>(uriBuilder.Uri, voucher);
         }
 
-        public async Task DisableVoucher(string code)
+        public IPromise DisableVoucher(string code)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder(string.Format("/vouchers/{0}/disable", code));
-            await this.client.DoPostRequest<dynamic>(uriBuilder.Uri);
+            return this.client.DoPostRequest(uriBuilder.Uri);
         } 
 
-        public async Task EnableVoucher(string code)
+        public IPromise EnableVoucher(string code)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder(string.Format("/vouchers/{0}/enable", code));
-            await this.client.DoPostRequest<dynamic>(uriBuilder.Uri);
+            return this.client.DoPostRequest(uriBuilder.Uri);
         }
 
-        public async Task<DataModel.VoucherRedemptionResult> Redeem(string code, DataModel.VoucherRedemptionQuery query)
+        public IPromise<DataModel.VoucherRedemptionResult> Redeem(string code, DataModel.VoucherRedemptionQuery query)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder(string.Format("/vouchers/{0}/redemption", code)).WithQuery(query);
-            return await this.client.DoPostRequest<DataModel.VoucherRedemptionResult>(uriBuilder.Uri);
+            return this.client.DoPostRequest<DataModel.VoucherRedemptionResult>(uriBuilder.Uri);
         }
 
-        public async Task<DataModel.VoucherRedemptionResult> Redeem(string code, DataModel.VoucherRedemptionContext context)
+        public IPromise<DataModel.VoucherRedemptionResult> Redeem(string code, DataModel.VoucherRedemptionContext context)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder(string.Format("/vouchers/{0}/redemption", code));
-            return await this.client.DoPostRequest<DataModel.VoucherRedemptionResult, DataModel.VoucherRedemptionContext>(new Uri(string.Format("/vouchers/{0}/redemption", code)), context);
+            return this.client.DoPostRequest<DataModel.VoucherRedemptionResult, DataModel.VoucherRedemptionContext>(new Uri(string.Format("/vouchers/{0}/redemption", code)), context);
         }
 
-        public async Task<DataModel.VoucherRedemption> Redemption(string code)
+        public IPromise<DataModel.VoucherRedemption> Redemption(string code)
         {
             UriBuilder uriBuilder = this.client.GetUriBuilder(string.Format("/vouchers/{0}/redemption", code));
-            return await this.client.DoGetRequest<DataModel.VoucherRedemption>(uriBuilder.Uri);
+            return this.client.DoGetRequest<DataModel.VoucherRedemption>(uriBuilder.Uri);
         }
     }
 }