+++ changelog.md
@@ -1,5 +1,6 @@
 Voucherify .NET SDK - Changelog
 ===============================
 
+- **2016-07-03** - `1.0.0` - Replace async/await with RSG.Promise library. Use RestSharp instead of HttpClient.
 - **2016-07-03** - `0.2.0` - Introduced support for .Net Framework 3.5.
 - **2016-06-26** - `0.1.0` - Inital verion of SDK that supports Vouchers, Redemption and Customer endpoints.